movehooks to separate file + login theming.
