+++ octoprint_mrbeam/static/css/loginui.css
@@ -1,0 +1,57 @@
+/* This lightweiht css file is only loaded for the loginui */
+/* See https://docs.octoprint.org/en/1.4.2/bundledplugins/loginui.html */
+
+a {
+  color: #e25303;
+}
+
+a:hover {
+  color: #a63e03;
+}
+
+a.muted,
+a.muted:hover {
+  color: #e25303a1;
+}
+
+button:focus {
+  outline: 0 !important;
+}
+
+.btn:focus,
+.btn:active,
+.btn.active {
+  box-shadow: none;
+}
+
+.btn-primary {
+  background: #e25303;
+  border-radius: 8px;
+  border-color: #e25303;
+  box-shadow: 1px 2px 6px rgba(0, 0, 0, 0.25);
+  color: #ffffff;
+  font-weight: 900;
+}
+
+.btn-primary:focus,
+.btn-primary:active,
+.btn-primary.active {
+  background: #e25303;
+  border: 2px solid #ffffff;
+  border-radius: 8px;
+  outline: none;
+}
+
+.btn-primary:hover {
+  background: #e25303;
+  box-shadow: 0px 1px 3px rgba(0, 0, 0, 0.25);
+}
+
+.btn-primary.disabled,
+.btn-primary[disabled] {
+  cursor: default;
+  background: #e4e4e4;
+  border-color: #e4e4e4;
+  box-shadow: none;
+  color: #383e42;
+}
